requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
Appium-Python-Client==2.7.1
Appium-UIAutomation==0.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
certifi==2022.9.24
exceptiongroup==1.0.4
h11==0.14.0
iniconfig==1.1.1
openssl-python==0.1.1
outcome==1.2.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
PyYAML==5.4.1
selenium==4.7.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
toml==0.10.2
urllib3==1.26.5
wsproto==1.2.0
